Prevent signals in threads from interfering
The threads we use in wire_io are very minimal and shouldn't be worried
about signals at all. We thus block them completely.

This helps when the main app uses signals to enable it to be the sole
receiver of signals and let it handle them where the logic to do so
really exists.