Automatically disable fd_state on no-op wait resume
If the fd_state was woken up but the resume was a no-op it means we are
now effectively in a busy-wait loop. Prevent that automatically by
disabling the fd_state.

This also means that we can skip explicitly disabling fd_state between
states and they will only get disabled if they are of a real impact on
the system. If the state transitioned automatically by the application
we can save ourselves some needless state switching.