Better fairness in wire_net
Do not let a single wire monopolize time when it has many consequtive
reads, let other wires get their time as well. It is fairly likely that
the reading/writing wire will actually take more time if it can't
complete in one shot so let the kernel buffers get ready for the next
operation instead of busy looping around it.