Optimize wire_io
reduce the number of system calls happening due to the wire io loop.

There was one system call since many times the fd was marked as pending
for read and then removed since there was nothing left to wait on, but
if there is nothing to read from the socketpair there is no wakeup as
well so we can just leave the wait in the kernel and reduce two system
calls.

Then there was the redundant read if the previous read didn't have a
full read. At worst we will increase latency slightly if there was a
request that was returned between the read and the end of the wire io
loop but that's always going to happen and reducing one mostly unlikely
to succeed read system call seems a better choice.